move rpc settings handling into separate file, rpc -> SuiRpcClient
